app/requirements.txt
fastapi
uvicorn
motor
pymongo
python-dotenv
certifi
dnspython
